refactor: 拆分 StorageService 的遷移掃描邏輯至 StorageMigrationScanner
- 新增 StorageMigrationScanner 類別，專責處理全量儲存空間掃描及資料遷移。
- 將 StorageService 中的 getAllHighlights 和 getAllSavedPageUrls 方法的邏輯移至 StorageMigrationScanner，降低 StorageService 的複雜度。
- 更新 migrationHandlers 以使用新的 migrationScanner 進行資料獲取，確保未來的擴展性和維護性。
- 移除 StorageService 中與遷移相關的過時方法，簡化其介面。
- 為 StorageMigrationScanner 添加單元測試，確保其功能正確性。
